fix(homepage): Adapt home page with app router
