+++ app/commands/page.tsx
@@ -1,3 +1,6 @@
+// TODO remove the use client directive in favor of a server component
+"use client";
+
 import { useState, useRef, useEffect } from "react";
 import commands from "../../data/commands.json";
 import Head from "next/head";
@@ -190,13 +193,13 @@ export default function Commands() {
             return null;
           })}
           {filteredCommands.length === 0 && (
-          <p className="select-none font-medium text-neutral-300">
-            No commands for your search were found!{" "}
-            <Link className="font-semibold" href="/discord" target="_blank">
-              Maybe suggest it?
-            </Link>
-          </p>
-        )}
+            <p className="select-none font-medium text-neutral-300">
+              No commands for your search were found!{" "}
+              <Link className="font-semibold" href="/discord" target="_blank">
+                Maybe suggest it?
+              </Link>
+            </p>
+          )}
         </div>
       </main>
     </>
