+++ components/Embeds/DailyMessageEmbed.tsx
@@ -10,45 +10,51 @@ import {
   DiscordThreadMessage,
 } from "@skyra/discord-components-react";
 import profiles from "../../data/profiles.json";
+import { useIsClient } from "@/helpers/hooks/useIsClient";
+import { FC } from "react";
 
 interface MainProps {
   threadName: string;
 }
 
-const DailyMessageEmbed: React.FC<MainProps> = ({ threadName }) => {
+const DailyMessageEmbed: FC<MainProps> = ({ threadName }) => {
+  const isClient = useIsClient();
+
   return (
-    <DiscordMessages className="min-w-fit overflow-x-hidden rounded-lg shadow">
-      <DiscordMessage
-        profile="wouldyou"
-        author={profiles.wouldyou.author}
-        avatar={profiles.wouldyou.avatar}
-        roleColor={profiles.wouldyou.roleColor}
-        bot={profiles.wouldyou.bot}
-        verified={profiles.wouldyou.verified}
-      >
-        <DiscordMention type="role" color="#1e99">
-          QOTD
-        </DiscordMention>
-        <DiscordEmbed slot="embeds" color="#1e88e5">
-          <DiscordEmbedDescription slot="description">
-            Would you rather be able to control fire 🔥 or water 💧?
-          </DiscordEmbedDescription>
-          <DiscordEmbedFooter slot="footer">
-            Daily Message | Type: Mixed | ID: 34
-          </DiscordEmbedFooter>
-        </DiscordEmbed>
-        <DiscordThread slot="thread" name={threadName}>
-          <DiscordThreadMessage
-            profile="Nightkiller"
-            author={profiles.nightkiller.author}
-            avatar={profiles.nightkiller.avatar}
-            roleColor={profiles.nightkiller.roleColor}
-          >
-            Wow that...
-          </DiscordThreadMessage>
-        </DiscordThread>
-      </DiscordMessage>
-    </DiscordMessages>
+    isClient && (
+      <DiscordMessages className="min-w-fit overflow-x-hidden rounded-lg shadow">
+        <DiscordMessage
+          profile="wouldyou"
+          author={profiles.wouldyou.author}
+          avatar={profiles.wouldyou.avatar}
+          roleColor={profiles.wouldyou.roleColor}
+          bot={profiles.wouldyou.bot}
+          verified={profiles.wouldyou.verified}
+        >
+          <DiscordMention type="role" color="#1e99">
+            QOTD
+          </DiscordMention>
+          <DiscordEmbed slot="embeds" color="#1e88e5">
+            <DiscordEmbedDescription slot="description">
+              Would you rather be able to control fire 🔥 or water 💧?
+            </DiscordEmbedDescription>
+            <DiscordEmbedFooter slot="footer">
+              Daily Message | Type: Mixed | ID: 34
+            </DiscordEmbedFooter>
+          </DiscordEmbed>
+          <DiscordThread slot="thread" name={threadName}>
+            <DiscordThreadMessage
+              profile="Nightkiller"
+              author={profiles.nightkiller.author}
+              avatar={profiles.nightkiller.avatar}
+              roleColor={profiles.nightkiller.roleColor}
+            >
+              Wow that...
+            </DiscordThreadMessage>
+          </DiscordThread>
+        </DiscordMessage>
+      </DiscordMessages>
+    )
   );
 };
 export default DailyMessageEmbed;
